requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --all-extras
#
attrs==25.3.0
    # via
    #   outcome
    #   trio
certifi==2025.4.26
    # via
    #   requests
    #   selenium
cffi==1.17.1
    # via trio
charset-normalizer==3.4.2
    # via requests
h11==0.16.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
numpy==2.3.0
    # via pandas
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
pandas==2.3.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.3
    # via -r requirements.in
selenium==4.34.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.14.1
    # via selenium
tzdata==2025.2
    # via pandas
urllib3[socks]==2.4.0
    # via
    #   requests
    #   selenium
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
